chore(config): remove angular from peerDependencies
Angular should be specified in peerDependencies, rather than
dependencies, because we want to ensure that the library consumer
has the right version.  However, it appears that the gulp task
generated by generator-ngx-library automatically moves the
"dependencies" specified in our package.json to "peerDependencies".
Thus we want to set our Angular peerDependencies as dependencies
so that we get the desired output on build without having to modify
the automatically generated gulp task.