feat(modal-window): split allowClose into more flexibile options
The previous allowClose functionality did too much: it both specified
whether clicking on the surrounding overlay would close the modal and
specified whether to show the built-in close button.  It's foreseeable and
entirely reasonable that a library consumer might want to disable the
close button while retaining the 'close the modal window on overlay
click' functionality.

In order to provide flexibility and allow for the above scenario, I'm
splitting allowClose into two distinct properties: closeOnOverlayClick
and showCloseButton.