refactor(demo): rearrange demo styles for clarity and explanation
I've moved some of the demo styles to a global stylesheet in order to
demonstrate the difference between in styling between global styles and
component styles.

In addition, I've moved my styling comments to the stylesheets and
reworded my blurb about ng-deep for clarity and to suggest using it
with caution.